Decrease waiting time for killing children.
Now, we give up after trying to kill children after 3 seconds.
This is because I can't figure out why `make` isn't being killed
by `cleanup`. It's desirable for killing to be faster so that
job cancellation doesn't take 15 seconds (only 3) if the children
can't be killed. We should address the root cause.